feat: rename MainHeader(DURI-340
