+++ packages/ui/src/components/Header/index.tsx
@@ -1,1 +1,1 @@
-export { Header } from './Header';
+export { MainHeader } from './MainHeader';
